manager/requirements.txt
absolufy-imports==0.3.1
annotated-types==0.7.0
anyio==4.10.0
apify_client==2.0.0
apify_shared==2.0.0
attrs==25.3.0
Authlib==1.6.1
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
cryptography==45.0.6
decorator==5.2.1
docstring_parser==0.17.0
fastapi==0.116.1
google-adk==1.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.106.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.35.1
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.28.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.3
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
impit==0.5.3
importlib_metadata==8.7.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
mcp==1.12.3
more-itertools==10.7.0
moviepy==2.2.1
numpy==2.3.2
opentelemetry-api==1.36.0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
pillow==11.3.0
proglog==0.1.12
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32==311
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
sse-starlette==3.0.2
starlette==0.47.2
tenacity==8.5.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
websockets==15.0.1
yt-dlp==2025.8.11
zipp==3.23.0
